pyproject.toml
[project]
name = "jira-bot"
version = "0.1.0"
description = "Telegram bot for creating Jira stories in AAI project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=20.0",
    "jira>=3.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "mypy>=1.0"
]
